pyproject.toml
[project]
name = "merchandiser-products-counter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.10.1",
    "faiss-cpu>=1.12.0",
    "fastapi[all]>=0.116.1",
    "from-root>=1.3.0",
    "ipykernel>=6.30.1",
    "numpy>=2.3.2",
    "open-clip-torch>=3.1.0",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "sentence-transformers>=5.1.0",
    "transformers>=4.56.1",
    "ultralytics>=8.3.191",
]
